+++ web/src/editor/inputs/ui-selector.tsx
@@ -2,10 +2,8 @@ import React, { useState, useEffect, useCallback } from 'react';
 import { ui_type, ui, ui_slider } from "shared/types/user-input";
 import { num_value, value_instance } from 'shared/types/mode';
 import { input, integer_input } from 'shared/types/parameters';
-// import { Slider } from '@fluentui/react';
 import { SliderConfig, default_slider_config } from '../../ui/slider';
 
-// import { value } from 'shared/types/parameters';
 import { Dropdown, IDropdownOption, IDropdownStyles } from '@fluentui/react';
 import { TextField, ITextFieldStyles } from '@fluentui/react/lib/TextField';
 import equal from "fast-deep-equal"
@@ -99,12 +97,8 @@ const Selector: React.FC<props> = ({ el, spec, path, onChange }) => {
             } catch (err) {
                 console.log("something went wrong when updating from the outside", err)
                 throw err
-
-
             }
         } else {
-            // console.log("Updating from the >> INSIDE <<")
-
             const val: ui | undefined = (
                 dropdown_key === "none" ?
                     undefined :
