+++ web/src/manual.tsx
@@ -1,5 +1,4 @@
-import React, { useEffect, useState } from "react"
-// DefaultButton, 
+import React, { useState } from "react"
 import { PrimaryButton } from '@fluentui/react/lib/Button';
 import { Slider } from '@fluentui/react/lib/Slider';
 import debounce from "lodash.debounce"
