+++ web/src/ui/button.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { DefaultButton } from '@fluentui/react/lib/Button';
-// import { button_value } from 'shared/types/mode';
 import { generic_ui } from 'shared/types/user-input';
 import { set_update } from "./utils"
